fix release not bottom space, remove old warning about available update.
