+++ packages/slice-machine/lib/consts.js
@@ -1,6 +1,5 @@
 export const warningStates = {
   NOT_CONNECTED: "NOT_CONNECTED",
-  NEW_VERSION_AVAILABLE: "NEW_VERSION_AVAILABLE",
   CLIENT_ERROR: "CLIENT_ERROR",
 };
